add base home screen & viewmodel
